WIP: apply aliasing and type filters to TypeScript inferences
