+++ src/mutations/codeFixes/extractRawCodefixUnionTypes.ts
@@ -2,7 +2,7 @@ import * as ts from "typescript";
 
 import { FileMutationsRequest } from "../../mutators/fileMutator";
 import { isNotUndefined } from "../../shared/arrays";
-import { findAliasOfType } from "../aliasing";
+import { findAliasOfType } from "../aliasing/findAliasOfType";
 
 /**
  * String equivalents to print for literal node types.
