+++ src/mutations/collecting.ts
@@ -4,7 +4,7 @@ import * as ts from "typescript";
 import { FileMutationsRequest } from "../mutators/fileMutator";
 import { setSubtract } from "../shared/sets";
 
-import { getApplicableTypeAliases } from "./aliasing";
+import { getApplicableTypeAliases } from "./aliasing/aliases";
 import { findMissingFlags, isTypeFlagSetRecursively } from "./collecting/flags";
 
 /**
