+++ src/mutations/generics.ts
@@ -3,7 +3,7 @@ import * as ts from "typescript";
 import { FileMutationsRequest } from "../mutators/fileMutator";
 import { isTypeBuiltIn } from "../shared/types";
 
-import { createTypeName } from "./aliasing";
+import { createTypeName } from "./aliasing/createTypeName";
 
 /**
  * Creates a type like "string[]" or "Map<boolean | number>" from a container and type arguments.
